@@ -1,6 +1,6 @@
 {
   "name": "fact-service",
-  "version": "1.1.1",
+  "version": "1.3.0",
   "type": "module",
   "description": "Enhanced Key-Value lookup service. Uses Nodejs, TypeScript, Express & Knex Project.",
   "main": "index.js",
@@ -13,7 +13,7 @@
   },
   "homepage": "https://github.com/justsml/fact-service",
   "scripts": {
-    "start": "nodemon --watch './**/*.ts' --exec 'ts-node' index.ts",
+    "start": "nodemon --watch './**/*.ts' --exec 'ts-node' ./servers/express/index.ts",
     "start:elysia": "bun run --watch ./servers/elysia/index.ts",
     "test": "vitest",
     "test:pg": "DB_ADAPTER=postgres vitest",
@@ -55,13 +55,13 @@
     "@elite-libs/auto-config": "^1.4.1",
     "@elysiajs/bearer": "^0.8.0",
     "@elysiajs/cors": "^0.8.0",
-    "@elysiajs/swagger": "^0.8.3",
+    "@elysiajs/swagger": "^0.8.5",
     "@google-cloud/firestore": "^7.1.0",
-    "axios": "^1.6.2",
+    "axios": "^1.6.8",
     "cassandra-driver": "^4.7.2",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
-    "elysia": "^0.8.9",
+    "elysia": "^0.8.17",
     "elysia-helmet": "^1.0.2",
     "express": "^4.18.2",
     "helmet": "^7.1.0",
@@ -70,28 +70,29 @@
     "lodash": "^4.17.21",
     "morgan": "^1.10.0",
     "pg": "^8.11.3",
-    "pino": "^8.17.1",
-    "pino-http": "^8.6.0",
-    "pino-pretty": "^10.3.0",
-    "ulidx": "^2.2.1"
+    "pino": "^8.19.0",
+    "pino-http": "^9.0.0",
+    "pino-pretty": "^10.3.1",
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "@faker-js/faker": "^8.3.1",
-    "@types/bun": "^1.0.0",
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.14",
-    "@types/lodash": "^4.14.189",
-    "@types/morgan": "^1.9.3",
-    "@types/ms": "^0.7.31",
-    "@types/node": "^20.11.0",
-    "@types/supertest": "^2.0.16",
-    "bun-types": "^1.0.20",
+    "@faker-js/faker": "^8.4.1",
+    "@types/bun": "^1.0.8",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/lodash": "^4.17.0",
+    "@types/morgan": "^1.9.9",
+    "@types/ms": "^0.7.34",
+    "@types/node": "^20.11.28",
+    "@types/supertest": "^6.0.2",
+    "@types/yargs": "^17.0.32",
+    "bun-types": "^1.0.31",
     "ms": "^2.1.3",
-    "nodemon": "^3.0.2",
-    "prettier": "^3.1.1",
-    "supertest": "^6.3.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.3.3",
-    "vitest": "^1.1.0"
+    "nodemon": "^3.1.0",
+    "prettier": "^3.2.5",
+    "supertest": "^6.3.4",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.2",
+    "vitest": "^1.4.0"
   }
 }
