@@ -201,7 +201,7 @@ function autoSetup<TError extends Error>(error: TError) {
     return setup()
       .catch((error) => logger.error(error))
       .then(() =>
-        Promise.reject(new UserError(`Table not found! Retry your request.`)),
+        Promise.reject(UserError(`Table not found! Retry your request.`)),
       );
   }
   throw error;
